backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2
xlrd==2.0.1
torch>=2.2.0
mariadb>=1.1.0
python-dotenv>=1.0.0
pyserial>=3.5
flask-socketio>=5.3.0

